Reflection

+++ src/portfoliotask4/Reflection05.java
@@ -8,8 +8,8 @@ public class Reflection05 {
 	    System.out.printf("There are %d fields\n", fields.length);
 
 	    for (Field f : fields) {
-	      System.out.printf("field name=%s type=%s value=%d\n", f.getName(),
-	          f.getType(), f.getInt(s));
+	      System.out.printf("field name=%s type=%s value=%f\n", f.getName(),
+	          f.getType(), f.getFloat(s));
 	    }
 	  }
 	}
