+++ src/portfoliotask4/Reflection10.java
@@ -5,7 +5,7 @@ import java.lang.reflect.Method;
 public class Reflection10 {
 	  public static void main(String[] args) throws Exception {
 		SimpleMultiply s = new SimpleMultiply();
-	    Method m = s.getClass().getDeclaredMethod("setA", int.class);
+	    Method m = s.getClass().getDeclaredMethod("setA", float.class);
 	    m.setAccessible(true);
 	    m.invoke(s, 76);
 	    System.out.println(s);
